order init

+++ database/migration/20250104061237_orders.up.sql
@@ -1,9 +1,9 @@
 CREATE TABLE orders
 (
     id          CHAR(36) PRIMARY KEY,
-    user_id     CHAR(36)       NOT NULL,
-    total_price DECIMAL(10, 2) NOT NULL,
+    user_id     CHAR(36) NOT NULL,
+    total_price int      NOT NULL,
     status      ENUM ('pending', 'processed', 'completed', 'canceled') DEFAULT 'pending',
-    created_at  TIMESTAMP                                              DEFAULT CURRENT_TIMESTAMP,
-    updated_at  TIMESTAMP                                              DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
+    created_at  TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+    updated_at  TIMESTAMP DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP
 );
